% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wlehar}
\alias{wlehar}
\title{Weight Lifting Exercises data from the Human Activity Recognition project.}
\format{A data frame with 39,242 observations of 159 variables.}
\source{
\url{http://groupware.les.inf.puc-rio.br/har#weight_lifting_exercises}
}
\usage{
data(wlehar)
}
\description{
A data set containing activity sensor measurements for six participants
performing dumbbell bicep curls.
}
\references{
\url{http://groupware.les.inf.puc-rio.br/work.jsf?p1=11201}
}
\keyword{datasets}

