% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wlehar_jhu_testing}
\alias{wlehar_jhu_testing}
\title{Weight Lifting Exercises testing data.}
\format{A data frame with 20 observations of 159 variables.}
\source{
Practical Machine Learning by John Hopkins University
 \url{https://d396qusza40orc.cloudfront.net/predmachlearn/pml-testing.csv}
}
\usage{
data(wlehar_jhu_testing)
}
\description{
Weight Lifting Exercises testing data.
}
\seealso{
\code{\link{wlehar}} for original data set.
}
\keyword{datasets}

