% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wlehar_jhu_training}
\alias{wlehar_jhu_training}
\title{Weight Lifting Exercises training data.}
\format{A data frame with 19,622 observations of 159 variables.}
\source{
Practical Machine Learning by John Hopkins University
 \url{https://d396qusza40orc.cloudfront.net/predmachlearn/pml-training.csv}
}
\usage{
data(wlehar_jhu_training)
}
\description{
Weight Lifting Exercises training data.
}
\seealso{
\code{\link{wlehar}} for original data set.
}
\keyword{datasets}

